Title: 합과 평균, Time: 8ms, Memory: 0MB, Status: Wrong Answer - Codetree

+++ 260101/합과 평균/sum-and-avg.cpp
@@ -3,14 +3,14 @@ using namespace std;
 
 int main() {
     // Please write your code here.
-    double A;
-    double B;
+    int A;
+    int B;
 
     cin >> A >> B;
 
-    double result = (A + B) / 2;
+    double result = (A + B) / 2.0;
 
     cout << A + B << " " << result;
-
+    
     return 0;
 }
